add text field validation on focus change
