+++ app/src/main/java/com/strv/movies/ui/login/Labels.kt
@@ -1,0 +1,21 @@
+package com.strv.movies.ui.login
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun Title() {
+    Text(
+        text = "Log In to Movies",
+        style = TextStyle(
+            color = MaterialTheme.colors.secondary,
+            fontWeight = FontWeight.Bold,
+            fontSize = 24.sp,
+            letterSpacing = 0.sp
+        )
+    )
+}
